implemented BA w/ reduction tree + density based patch selector (from RAMPVO))

+++ devo/config.py
@@ -31,6 +31,6 @@ _C.MOTION_DAMPING = 0.5
 
 _C.MIXED_PRECISION = True
 
-_C.BA_PRECISION = 'double'  # 'float32', 'double', 'float32'
+_C.BA_PRECISION = 'float32'  # 'float32', 'double', 'float32'
 
 cfg = _C
